basic-chat/requirements.txt
boto3
streamlit
pyyaml